requirements.txt
Pillow==10.3.0
geopy==2.4.1
staticmap==0.5.4